requirements.txt
requests==2.31.0
paypalrestsdk==1.13.1
mercadopago==2.2.7
python-telegram-bot==20.3
python-dotenv==1.0.0
Flask==2.2.3
python-dateutil==2.8.2
cryptography==41.0.4